refactor: Order 금액 필드 타입을 Long으로 변경
설계문서(ERD)와의 정합성을 위해 Order, OrderProduct의 금액 관련 필드들을
int에서 Long으로 변경합니다.

- Order: totalAmount, shippingFee, discountAmount, paymentAmount
- OrderProduct: price, extraPrice, calculateTotalPrice() 반환 타입

관련 DTO(Info, Response)들과 테스트 코드도 함께 수정되었습니다.

+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderAdminDetailInfo.java
@@ -17,10 +17,10 @@ public record OrderAdminDetailInfo(
     String address,
     String addressDetail,
     String shippingMemo,
-    int totalAmount,
-    int shippingFee,
-    int discountAmount,
-    int paymentAmount,
+    Long totalAmount,
+    Long shippingFee,
+    Long discountAmount,
+    Long paymentAmount,
     List<OrderProductInfo> orderProducts
 ) {
