+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/order/OrderAdminV1Dto.java
@@ -15,8 +15,8 @@ public class OrderAdminV1Dto {
         String orderNumber,
         String orderName,
         OrderStatus status,
-        int totalAmount,
-        int paymentAmount,
+        Long totalAmount,
+        Long paymentAmount,
         String thumbnailUrl
     ) {
         public static OrderAdminResponse from(OrderInfo info) {
@@ -44,10 +44,10 @@ public class OrderAdminV1Dto {
         String address,
         String addressDetail,
         String shippingMemo,
-        int totalAmount,
-        int shippingFee,
-        int discountAmount,
-        int paymentAmount,
+        Long totalAmount,
+        Long shippingFee,
+        Long discountAmount,
+        Long paymentAmount,
         List<OrderProductResponse> orderProducts
     ) {
         public static OrderAdminDetailResponse from(OrderAdminDetailInfo info) {
@@ -81,8 +81,8 @@ public class OrderAdminV1Dto {
         Long productOptionId,
         String productName,
         String optionName,
-        int price,
-        int extraPrice,
+        Long price,
+        Long extraPrice,
         int quantity,
         String thumbnailUrl,
         OrderProductStatus status
